+++ pyatoa/tests/test_config.py
@@ -8,13 +8,6 @@ from pyasdf import ASDFDataSet
 from pyatoa.core.config import Config
 
 
-def test_read_config_from_seisflow_yaml():
-    """
-    Test that reading from an external Seisflows YAML file works
-    """
-    cfg = Config(seisflows_yaml="./test_data/test_seisflows_parameters.yaml")
-    assert(cfg.synthetics_only == True)  # Check a random variable
-
 
 def test_io_asdf(tmpdir):
     """